==> R (mouse)/10-na12878_analysis.R <==
##########################################################
######         Workflow Part 10: NA878 cells       #####
##########################################################
library(dplyr)
library(GenomicRanges)
library(tidyr)
library(ggplot2)
library(breakpointR)


######################## Collect Data ######################

# Datasets
quality = read.table("../NA878/quality.lansdorp_na12878.txt",header=F)
quality_libraries = filter(quality, V2>0.7)$V1

collectBreaksAllFiles <- function(datapath="../StraVa/DATA/rdata/"){
  files <- datapath
  
  
  breaks.all.files <- data.frame()
  n=1
  for (file in files) {
    message("Reading ... " , basename(file), " ... ",round(  (n/length(files))*100  ,  digits = 1  ) , "%"  )
    n=n+1
    data <- get(load(file))[c('breaks','ID')]
    breakpoints <- as.data.frame(data$breaks)
    
    
    if (nrow(breakpoints)) {
      breakpoints$library = data$ID
      breaks.all.files <- rbind(breakpoints,breaks.all.files)
    }  
  }
  return(breaks.all.files)
}

na878_bp = collectBreaksAllFiles(datapath = paste0("../NA878/data/",quality_libraries,".RData")) # 59K

# Merge list of good libraries
na878_breakpoints= merge(na878_bp, quality, by.x="library",by.y="V1") # 28k 
na878_breakpoints$library <- as.factor(na878_breakpoints$library)
na878_breakpoints_good = filter(na878_breakpoints, V2 > 0.7) # 28k 
length(levels(na878_breakpoints_good$library))

######################## Filtering data ######################

# 1) Remove homozygous events
na878_breakpoints_good_hetero = filter(na878_breakpoints_good,genoT!= "cc-ww" | genoT!= "cc-ww") # 26k 
#levels(na878_breakpoints_good_hetero$library) <- factor(na878_breakpoints_good_hetero$library)
#length(levels(na878_breakpoints_good_hetero$library))
na878_breakpoints_good_hetero$library<-as.factor(na878_breakpoints_good_hetero$library)
# 2) Filter out events that are too close to each other (2Mb)
na878_breakpoints_good_hetero_SCE = data.frame()
n=1
level= levels(na878_breakpoints_good_hetero$library)[1]
for (level in levels(na878_breakpoints_good_hetero$library)){
  message(  round(     (      (n/length(levels(na878_breakpoints_good_hetero$library)))   *100        )     ,digits = 2)   ,"% ... complete"   )
  tmp = filter(na878_breakpoints_good_hetero, library==level)
  tmp$seqnames <- droplevels(tmp$seqnames)
  
  level2=levels(tmp$seqnames)[1]
  for (level2 in levels(tmp$seqnames)){
    
    tmp2 = filter(tmp, seqnames==level2)
    tmp3 <- GRanges(tmp2)
    
    overlaps = countOverlaps(tmp3, tmp3, type="any",maxgap = 2500000)
    overlap_rows = which(overlaps %in% c(1))
    
    tmp4 = tmp2[overlap_rows,]

    na878_breakpoints_good_hetero_SCE <- rbind(na878_breakpoints_good_hetero_SCE,tmp4)
    
  }
  n=n+1
}
na878_breakpoints_good_hetero_SCE$library<-droplevels(na878_breakpoints_good_hetero_SCE$library)
length(levels(na878_breakpoints_good_hetero_SCE$library))

# 3) Filter out events near centromeres
centromeres <- read.table("centromeres2.txt",header=F,fill=T) #%>% select(-c(V4))
centromeres <-centromeres %>% dplyr::rename("seqnames"=V1,"start"=V2,"end"=V3)
centroGRange <- GRanges(centromeres)
summaryBreaks.df <- GRanges(na878_breakpoints_good_hetero_SCE)
na878_breakpoints_good_hetero_SCE_CM_bl <-as.data.frame(summaryBreaks.df[-queryHits(findOverlaps(summaryBreaks.df, centroGRange, type="any")),])
na878_breakpoints_good_hetero_SCE_CM_bl$library<-droplevels(na878_breakpoints_good_hetero_SCE_CM_bl$library)

# 4) Filter out events that are too close to each other (2Mb)
na878_breakpoints_good_hetero_SCE_CM_bl_filtered_10kb <- filter(na878_breakpoints_good_hetero_SCE_CM_bl, width < 10000) # 3k 
na878_breakpoints_good_hetero_SCE_CM_bl_filtered_10kb$library<-droplevels(na878_breakpoints_good_hetero_SCE_CM_bl_filtered_10kb$library)
                                      

######################## Exporting Data ######################

# Write to enrichment folder for enrichment analysis
write.table(select(na878_breakpoints_good_hetero_SCE_CM_bl,c(seqnames, start,end)), "Thesis/Enrichment/GenomePermute/na878_sces.bed", col.names = F, row.names = F, quote = F, sep="\t")

sces=read.table("Thesis/Enrichment/GenomePermute/na878_sces.bed")
sces$width=sces$V3-sces$V2
ggplot(sces)+geom_density(aes(width)) + scale_x_log10()
median(sces$width)


######################## Plotting Data ######################
# Take putative SCEs and reprint breaksPlot with just them




cleanRData <- function(data=na878_breakpoints_good_hetero_SCE_CM_bl){
  n=1
  for (file in levels(data$library)){
    
    message(  round(     (      (n/length(levels(data$library)))   *100        )     ,digits = 2)   ,"% ... complete"   )
    filename = paste0("../NA878/data/",file,".RData")
    tmp = get(load(filename))
    
    seqinfo <- tmp$breaks@seqinfo
    seqnameLevels <- levels(tmp$breaks@seqnames)
    
    tmp$breaks <-GRanges(filter(data,library==tmp$ID) %>% select(c(seqnames,start,end,width,strand,genoT,deltaW)))
    tmp$breaks@seqinfo <- seqinfo
    levels(tmp$breaks@seqnames) <- seqnameLevels
    
    tmp$confint = tmp$confint[queryHits(findOverlaps(tmp$confint, tmp$breaks, type="any")),]
    #$confint<-GRanges(filter(breaks,library==tmp$ID) %>% select(c(seqnames,CI.start,CI.end,width,strand,genoT,deltaW)))
    tmp$confint@seqinfo <- seqinfo
    levels(tmp$confint@seqnames) <- seqnameLevels
    
    save(tmp, file=paste0("../NA878/data_filtered/",file,"filtered.RData"))
    
    n=n+1
  }
  
}

cleanRData(na878_breakpoints_good_hetero)
breakpointR::plotBreakpoints(files2plot = list.files("../NA878/data_filtered/",full.names = T),file = "breaksPlot_na12878_good7_hetero.pdf")

cleanRData(na878_breakpoints_good_hetero_SCE)
breakpointR::plotBreakpoints(files2plot = list.files("../NA878/data_filtered/",full.names = T),file = "breaksPlot_na12878_good7_hetero_sce.pdf")

cleanRData(na878_breakpoints_good_hetero_SCE_CM_bl)
breakpointR::plotBreakpoints(files2plot = list.files("../NA878/data_filtered/",full.names = T),file = "breaksPlot_na12878_good7_hetero_sce_cm.pdf")

cleanRData(na878_breakpoints_good_hetero_SCE_CM_bl_filtered_10kb)
breakpointR::plotBreakpoints(files2plot = list.files("../NA878/data_filtered/",full.names = T),file = "breaksPlot_na12878_good7_SCE_CM_10KB.pdf")
